+++ relay-server/src/metrics_extraction/generic.rs
@@ -218,6 +218,7 @@ mod tests {
                     received_at: Some(
                         UnixTimestamp(0),
                     ),
+                    extracted_from_indexed: false,
                 },
             },
         ]
@@ -268,6 +269,7 @@ mod tests {
                     received_at: Some(
                         UnixTimestamp(0),
                     ),
+                    extracted_from_indexed: false,
                 },
             },
         ]
@@ -320,6 +322,7 @@ mod tests {
                     received_at: Some(
                         UnixTimestamp(0),
                     ),
+                    extracted_from_indexed: false,
                 },
             },
         ]
@@ -386,6 +389,7 @@ mod tests {
                     received_at: Some(
                         UnixTimestamp(0),
                     ),
+                    extracted_from_indexed: false,
                 },
             },
         ]
@@ -449,6 +453,7 @@ mod tests {
                     received_at: Some(
                         UnixTimestamp(0),
                     ),
+                    extracted_from_indexed: false,
                 },
             },
         ]
@@ -516,6 +521,7 @@ mod tests {
                     received_at: Some(
                         UnixTimestamp(0),
                     ),
+                    extracted_from_indexed: false,
                 },
             },
         ]
@@ -591,6 +597,7 @@ mod tests {
                     received_at: Some(
                         UnixTimestamp(0),
                     ),
+                    extracted_from_indexed: false,
                 },
             },
         ]